--- src/main/resources/view/HelpMenu.fxml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.net.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.image.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.stage.*?>
-
-<fx:root resizable="false" title="Help" type="javafx.stage.Stage" xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1">
-    <icons>
-        <Image url="@/images/help_icon.png" />
-    </icons>
-    <scene>
-        <Scene>
-            <stylesheets>
-                <URL value="@HelpWindow.css" />
-            </stylesheets>
-         <VBox>
-            <children>
-                <StackPane>
-                    <GridPane fx:id="page1" hgap="1" prefWidth="1000.0" style="-fx-background-color: #383838;" vgap="1">
-                        <columnConstraints>
-                            <ColumnConstraints halignment="LEFT" hgrow="SOMETIMES" maxWidth="375.0" minWidth="10.0" prefWidth="375.0" />
-                            <ColumnConstraints halignment="LEFT" hgrow="SOMETIMES" maxWidth="500.0" minWidth="10.0" prefWidth="500.0" />
-                            <ColumnConstraints halignment="LEFT" hgrow="SOMETIMES" maxWidth="500.0" minWidth="10.0" prefWidth="500.0" />
-                        </columnConstraints>
-                        <children>
-                        </children>
-                    </GridPane>
-                    <GridPane fx:id="page2" hgap="1" prefWidth="1000.0" style="-fx-background-color: #383838;" vgap="1">
-                        <columnConstraints>
-                            <ColumnConstraints halignment="LEFT" hgrow="SOMETIMES" maxWidth="375.0" minWidth="10.0" prefWidth="375.0" />
-                            <ColumnConstraints halignment="LEFT" hgrow="SOMETIMES" maxWidth="500.0" minWidth="10.0" prefWidth="500.0" />
-                            <ColumnConstraints halignment="LEFT" hgrow="SOMETIMES" maxWidth="500.0" minWidth="10.0" prefWidth="500.0" />
-                        </columnConstraints>
-                        <children>
-                        </children>
-                    </GridPane>
-                </StackPane>
-               <HBox alignment="CENTER" prefHeight="0.0" prefWidth="1000.0" style="-fx-background-color: #383838;">
-                  <children>
-                     <Button fx:id="firstPageBtn" styleClass="button" mnemonicParsing="false" onAction="#setPage1" text="1" />
-                     <Button fx:id="secondPageBtn" styleClass="button" mnemonicParsing="false" onAction="#setPage2" text="2" />
-                  </children>
-               </HBox>
-            </children>
-         </VBox>
-        </Scene>
-    </scene>
-</fx:root>
